Fix stoichiometry inventory-update dialog race condition
The Update Inventory Stock dialog computed its default molecule
selection from linkedInventoryQuantityInfoByGlobalId, but that map is
populated asynchronously by per-molecule inventory quantity queries.
If the dialog was opened before all queries resolved, or before an
open->close->open cycle finished resetting state via a ref-tracked
effect, it could preselect molecules based on incomplete or stale
quantity data.

- Split the dialog into a thin shell (StoichiometryInventoryUpdateDialog)
  and a stateful content component (StoichiometryInventoryUpdateDialogContent).
  All state now lives in the content component and is initialized once
  via a useState initializer; since MUI's Dialog unmounts its children on
  close, each open is a fresh mount with no reset effects/refs needed.
- Gate the "Update Inventory Stock" toolbar button on all linked
  molecules' inventory quantities having resolved
  (StoichiometryTableToolbar), so the dialog can no longer open with an
  incomplete quantity map.
- Fix useSubSampleQuantitiesQuery so a query that throws (network
  failure, unparseable response) also lands an error entry in the
  returned map, matching the existing !response.ok handling. Without
  this, a thrown query error left the molecule out of the map entirely
  and the toolbar's loading check never resolved, permanently disabling
  the button.
- Drop the incidental React.lazy/Suspense wrapper added around the
  dialog content - the mount/unmount split is what fixes the race, the
  lazy loading was unrelated overhead.
