+++ src/main/webapp/ui/src/tinyMCE/stoichiometry/StoichiometryInventoryUpdateDialog.tsx
@@ -1,39 +1,13 @@
-import WarningAmberIcon from "@mui/icons-material/WarningAmber";
-import Alert from "@mui/material/Alert";
-import AlertTitle from "@mui/material/AlertTitle";
-import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogTitle from "@mui/material/DialogTitle";
-import Paper from "@mui/material/Paper";
-import Stack from "@mui/material/Stack";
-import Table from "@mui/material/Table";
-import TableBody from "@mui/material/TableBody";
-import TableCell from "@mui/material/TableCell";
-import TableContainer from "@mui/material/TableContainer";
-import TableHead from "@mui/material/TableHead";
-import TableRow from "@mui/material/TableRow";
-import Typography from "@mui/material/Typography";
-import { useMutation } from "@tanstack/react-query";
 import React from "react";
 import { Dialog } from "@/components/DialogBoundary";
 import type { InventoryQuantityQueryResult } from "@/modules/inventory/queries";
-import StoichiometryInventoryUpdateMoleculeRow from "@/tinyMCE/stoichiometry/StoichiometryInventoryUpdateMoleculeRow";
+import StoichiometryInventoryUpdateDialogContent, {
+  type InventoryStockUpdateResult,
+} from "@/tinyMCE/stoichiometry/StoichiometryInventoryUpdateDialogContent";
 import type { EditableMolecule } from "@/tinyMCE/stoichiometry/types";
-import { getInventoryUpdateEligibility } from "@/tinyMCE/stoichiometry/utils";
-export type InventoryStockUpdateResult = {
-  refreshedStoichiometry?: {
-    id: number;
-    revision: number;
-  };
-  results: ReadonlyArray<{
-    moleculeId: number;
-    moleculeName: string;
-    success: boolean;
-    errorMessage: string | null;
-  }>;
-};
+
+export type { InventoryStockUpdateResult };
+
 type StoichiometryInventoryUpdateDialogProps = {
   open: boolean;
   molecules: ReadonlyArray<EditableMolecule>;
@@ -41,17 +15,12 @@ type StoichiometryInventoryUpdateDialogProps = {
   onSave?: (selectedMoleculeIds: number[]) => Promise<InventoryStockUpdateResult>;
   onClose: () => void;
 };
-function getDefaultValues(
-  molecules: ReadonlyArray<EditableMolecule>,
-  linkedInventoryQuantityInfoByGlobalId: ReadonlyMap<string, InventoryQuantityQueryResult>,
-): number[] {
-  return molecules
-    .filter((molecule) => {
-      const eligibility = getInventoryUpdateEligibility(molecule, linkedInventoryQuantityInfoByGlobalId);
-      return eligibility.disabledReason === null && molecule.inventoryLink?.stockDeducted !== true;
-    })
-    .map(({ id }) => id);
-}
+
+/**
+ * Thin shell around the Update Inventory Stock dialog. All state lives in the
+ * content component, which MUI mounts fresh on each open and unmounts on
+ * close, so opening the dialog is what resets it.
+ */
 export default function StoichiometryInventoryUpdateDialog({
   open,
   molecules,
@@ -60,95 +29,10 @@ export default function StoichiometryInventoryUpdateDialog({
   onClose,
 }: StoichiometryInventoryUpdateDialogProps): React.ReactNode {
   const titleId = React.useId();
-  const wasOpenRef = React.useRef(false);
-  const [selectedMoleculeIds, setSelectedMoleculeIds] = React.useState<number[]>([]);
-  const saveMutation = useMutation({
-    mutationFn: async (moleculeIds: number[]) => {
-      if (!onSave) {
-        return {
-          results: [],
-        } satisfies InventoryStockUpdateResult;
-      }
-      return onSave(moleculeIds);
-    },
-  });
-  const failedResults = saveMutation.data?.results.filter(({ success }) => !success) ?? [];
-  const saveFeedback =
-    saveMutation.isSuccess && failedResults.length > 0
-      ? "Current stock amounts were refreshed. Re-select any remaining molecules to retry."
-      : null;
-  const saveError = (() => {
-    if (saveMutation.isError) {
-      const message = saveMutation.error.message;
-      return `${message} Current stock amounts were refreshed where possible.`;
-    }
-    if (failedResults.length === 0) {
-      return null;
-    }
-    return (
-      Array.from(
-        new Set(
-          failedResults
-            .map(({ errorMessage }) => errorMessage)
-            .filter((message): message is string => Boolean(message)),
-        ),
-      ).join(" ") || null
-    );
-  })();
-  const hasInvalidSelectedRows = selectedMoleculeIds.some((selectedMoleculeId) => {
-    const selectedMolecule = molecules.find(({ id }) => id === selectedMoleculeId);
-    if (!selectedMolecule) {
-      return true;
-    }
-    return (
-      getInventoryUpdateEligibility(selectedMolecule, linkedInventoryQuantityInfoByGlobalId).disabledReason !== null
-    );
-  });
-  const selectionError = hasInvalidSelectedRows ? "Re-select any invalid molecules before saving." : null;
-  const resetDialogState = React.useCallback(() => {
-    setSelectedMoleculeIds(getDefaultValues(molecules, linkedInventoryQuantityInfoByGlobalId));
-    saveMutation.reset();
-  }, [linkedInventoryQuantityInfoByGlobalId, molecules, saveMutation]);
-  React.useEffect(() => {
-    if (open && !wasOpenRef.current) {
-      resetDialogState();
-    }
-    wasOpenRef.current = open;
-  }, [open, resetDialogState]);
-  const handleClose = React.useCallback(() => {
-    resetDialogState();
-    onClose();
-  }, [onClose, resetDialogState]);
-  const handleSubmit = React.useCallback(async () => {
-    if (selectedMoleculeIds.length === 0) {
-      saveMutation.reset();
-      return;
-    }
-    if (hasInvalidSelectedRows) {
-      saveMutation.reset();
-      return;
-    }
-    if (!onSave) {
-      handleClose();
-      return;
-    }
-    saveMutation.reset();
-    try {
-      const result = await saveMutation.mutateAsync(selectedMoleculeIds);
-      const hasFailures = result.results.some(({ success }) => !success);
-      if (!hasFailures) {
-        handleClose();
-        return;
-      }
-      setSelectedMoleculeIds([]);
-    } catch {
-      setSelectedMoleculeIds([]);
-    }
-  }, [handleClose, hasInvalidSelectedRows, onSave, saveMutation, selectedMoleculeIds]);
   return (
     <Dialog
       open={open}
-      onClose={handleClose}
+      onClose={onClose}
       maxWidth="lg"
       fullWidth
       aria-labelledby={titleId}
@@ -160,128 +44,13 @@ export default function StoichiometryInventoryUpdateDialog({
         },
       }}
     >
-      <Box
-        component="form"
-        sx={{
-          display: "contents",
-        }}
-        onSubmit={(event) => {
-          event.preventDefault();
-          event.stopPropagation();
-          void handleSubmit();
-        }}
-      >
-        <DialogTitle id={titleId}>Update Inventory Stock</DialogTitle>
-        <DialogContent>
-          <Stack spacing={2}>
-            <Typography variant="body2" color="text.secondary">
-              Select the molecules from this stoichiometry table whose linked inventory stock should be updated.
-            </Typography>
-            <Alert
-              severity="warning"
-              icon={
-                <WarningAmberIcon
-                  aria-label="Action irreversible warning"
-                  fontSize="small"
-                  sx={{
-                    color: "warning.main",
-                  }}
-                />
-              }
-            >
-              <AlertTitle>WARNING: This action is irreversible</AlertTitle>
-
-              <Typography variant="body2" gutterBottom>
-                <strong>This will permanently reduce inventory quantities.</strong>
-              </Typography>
-              <Typography variant="body2" gutterBottom>
-                Stock cannot be automatically replenished if you change quantities later, delete this stoichiometry
-                table, delete the document, or unlink samples.
-              </Typography>
-              <Typography variant="body2">
-                Only proceed if you have actually used these materials in your experiment.
-              </Typography>
-            </Alert>
-            {selectionError && <Alert severity="warning">{selectionError}</Alert>}
-            {saveFeedback && <Alert severity="info">{saveFeedback}</Alert>}
-            {saveError && <Alert severity="error">{saveError}</Alert>}
-            <TableContainer
-              component={Paper}
-              variant="outlined"
-              sx={{
-                maxHeight: 560,
-                overflowY: "auto",
-                borderRadius: 1,
-              }}
-            >
-              <Table stickyHeader size="small">
-                <TableHead>
-                  <TableRow>
-                    <TableCell
-                      padding="checkbox"
-                      aria-label="Select molecule"
-                      width={52}
-                      sx={{
-                        px: 0.5,
-                      }}
-                    />
-                    <TableCell width="55%">Molecule</TableCell>
-                    <TableCell align="right" width="15%">
-                      In Stock
-                    </TableCell>
-                    <TableCell align="right" width="15%">
-                      Will Use
-                    </TableCell>
-                    <TableCell align="right" width="15%">
-                      Remaining
-                    </TableCell>
-                  </TableRow>
-                </TableHead>
-                <TableBody>
-                  {molecules.map((molecule) => {
-                    const eligibility = getInventoryUpdateEligibility(molecule, linkedInventoryQuantityInfoByGlobalId);
-                    const { disabledReason, helperText, stockDisplay } = eligibility;
-                    const disabled = disabledReason !== null;
-                    const selected = !disabled && selectedMoleculeIds.includes(molecule.id);
-                    return (
-                      <StoichiometryInventoryUpdateMoleculeRow
-                        key={molecule.id}
-                        molecule={molecule}
-                        selected={selected}
-                        disabled={disabled}
-                        helperText={helperText}
-                        stockDisplay={stockDisplay}
-                        onToggle={() => {
-                          if (disabled) {
-                            return;
-                          }
-                          saveMutation.reset();
-                          setSelectedMoleculeIds((currentIds) =>
-                            selected ? currentIds.filter((id) => id !== molecule.id) : [...currentIds, molecule.id],
-                          );
-                        }}
-                      />
-                    );
-                  })}
-                </TableBody>
-              </Table>
-            </TableContainer>
-          </Stack>
-        </DialogContent>
-        <DialogActions>
-          <Button type="button" onClick={handleClose}>
-            Cancel
-          </Button>
-          <Button
-            type="submit"
-            variant="contained"
-            color="callToAction"
-            disabled={selectedMoleculeIds.length === 0 || hasInvalidSelectedRows || saveMutation.isPending}
-          >
-            {saveMutation.isPending ? "Saving..." : "Save"}
-          </Button>
-        </DialogActions>
-      </Box>
+      <StoichiometryInventoryUpdateDialogContent
+        titleId={titleId}
+        molecules={molecules}
+        linkedInventoryQuantityInfoByGlobalId={linkedInventoryQuantityInfoByGlobalId}
+        onSave={onSave}
+        onClose={onClose}
+      />
     </Dialog>
   );
 }
